+++ lib/card_settings.dart
@@ -38,3 +38,4 @@ export 'widgets/text_fields/card_settings_text.dart';
 export 'helpers/picker_dialog.dart';
 export 'helpers/scroll_picker.dart';
 export 'helpers/converter_functions.dart';
+export 'helpers/layout_functions.dart';